+++ compiler/ml/ast_iterator.ml
@@ -96,9 +96,9 @@ module T = struct
     sub.attributes sub attrs;
     match desc with
     | Ptyp_any | Ptyp_var _ -> ()
-    | Ptyp_arrow (_lab, t1, t2, _) ->
-      sub.typ sub t1;
-      sub.typ sub t2
+    | Ptyp_arrow {arg; ret} ->
+      sub.typ sub arg;
+      sub.typ sub ret
     | Ptyp_tuple tyl -> List.iter (sub.typ sub) tyl
     | Ptyp_constr (lid, tl) ->
       iter_loc sub lid;