+++ compiler/ml/depend.ml
@@ -105,9 +105,9 @@ let rec add_type bv ty =
   match ty.ptyp_desc with
   | Ptyp_any -> ()
   | Ptyp_var _ -> ()
-  | Ptyp_arrow (_, t1, t2, _) ->
-    add_type bv t1;
-    add_type bv t2
+  | Ptyp_arrow {arg; ret} ->
+    add_type bv arg;
+    add_type bv ret
   | Ptyp_tuple tl -> List.iter (add_type bv) tl
   | Ptyp_constr (c, tl) ->
     add bv c;