+++ compiler/syntax/src/jsx_v4.ml
@@ -1305,16 +1305,14 @@ let transform_structure_item ~config item =
       let rec get_prop_types types
           ({ptyp_loc; ptyp_desc; ptyp_attributes} as full_type) =
         match ptyp_desc with
-        | Ptyp_arrow (name, type_, ({ptyp_desc = Ptyp_arrow _} as rest), _)
+        | Ptyp_arrow {lbl = name; arg; ret = {ptyp_desc = Ptyp_arrow _} as typ2}
           when is_labelled name || is_optional name ->
-          get_prop_types
-            ((name, ptyp_attributes, ptyp_loc, type_) :: types)
-            rest
-        | Ptyp_arrow (Nolabel, _type, rest, _) -> get_prop_types types rest
-        | Ptyp_arrow (name, type_, return_value, _)
+          get_prop_types ((name, ptyp_attributes, ptyp_loc, arg) :: types) typ2
+        | Ptyp_arrow {lbl = Nolabel; ret} -> get_prop_types types ret
+        | Ptyp_arrow {lbl = name; arg; ret = return_value}
           when is_labelled name || is_optional name ->
           ( return_value,
-            (name, ptyp_attributes, return_value.ptyp_loc, type_) :: types )
+            (name, ptyp_attributes, return_value.ptyp_loc, arg) :: types )
         | _ -> (full_type, types)
       in
       let inner_type, prop_types = get_prop_types [] pval_type in
@@ -1409,21 +1407,27 @@ let transform_signature_item ~config item =
       let rec get_prop_types types ({ptyp_loc; ptyp_desc} as full_type) =
         match ptyp_desc with
         | Ptyp_arrow
-            ( name,
-              ({ptyp_attributes = attrs} as type_),
-              ({ptyp_desc = Ptyp_arrow _} as rest),
-              _ )
-          when is_optional name || is_labelled name ->
-          get_prop_types ((name, attrs, ptyp_loc, type_) :: types) rest
+            {
+              lbl;
+              arg = {ptyp_attributes = attrs} as type_;
+              ret = {ptyp_desc = Ptyp_arrow _} as rest;
+            }
+          when is_optional lbl || is_labelled lbl ->
+          get_prop_types ((lbl, attrs, ptyp_loc, type_) :: types) rest
         | Ptyp_arrow
-            ( Nolabel,
-              {ptyp_desc = Ptyp_constr ({txt = Lident "unit"}, _)},
-              rest,
-              _ ) ->
+            {
+              lbl = Nolabel;
+              arg = {ptyp_desc = Ptyp_constr ({txt = Lident "unit"}, _)};
+              ret = rest;
+            } ->
           get_prop_types types rest
-        | Ptyp_arrow (Nolabel, _type, rest, _) -> get_prop_types types rest
+        | Ptyp_arrow {lbl = Nolabel; ret = rest} -> get_prop_types types rest
         | Ptyp_arrow
-            (name, ({ptyp_attributes = attrs} as type_), return_value, _)
+            {
+              lbl = name;
+              arg = {ptyp_attributes = attrs} as type_;
+              ret = return_value;
+            }
           when is_optional name || is_labelled name ->
           (return_value, (name, attrs, return_value.ptyp_loc, type_) :: types)
         | _ -> (full_type, types)