+++ compiler/syntax/src/res_parsetree_viewer.ml
@@ -7,30 +7,28 @@ let arrow_type ?(max_arity = max_int) ct =
   let rec process attrs_before acc typ arity =
     match typ with
     | _ when arity < 0 -> (attrs_before, List.rev acc, typ)
-    | {ptyp_desc = Ptyp_arrow (_, _, _, Some _); ptyp_attributes = []}
+    | {ptyp_desc = Ptyp_arrow {arity = Some _}; ptyp_attributes = []}
       when acc <> [] ->
       (attrs_before, List.rev acc, typ)
     | {
-     ptyp_desc = Ptyp_arrow ((Nolabel as lbl), typ1, typ2, _);
+     ptyp_desc = Ptyp_arrow {lbl = Nolabel as lbl; arg; ret};
      ptyp_attributes = [];
     } ->
-      let arg = ([], lbl, typ1) in
-      process attrs_before (arg :: acc) typ2 (arity - 1)
+      let arg = ([], lbl, arg) in
+      process attrs_before (arg :: acc) ret (arity - 1)
     | {
-     ptyp_desc = Ptyp_arrow (Nolabel, _typ1, _typ2, _);
+     ptyp_desc = Ptyp_arrow {lbl = Nolabel};
      ptyp_attributes = [({txt = "bs"}, _)];
     } ->
       (* stop here, the uncurried attribute always indicates the beginning of an arrow function
          * e.g. `(. int) => (. int)` instead of `(. int, . int)` *)
       (attrs_before, List.rev acc, typ)
-    | {
-        ptyp_desc = Ptyp_arrow (Nolabel, _typ1, _typ2, _);
-        ptyp_attributes = _attrs;
-      } as return_type ->
+    | {ptyp_desc = Ptyp_arrow {lbl = Nolabel}; ptyp_attributes = _attrs} as
+      return_type ->
       let args = List.rev acc in
       (attrs_before, args, return_type)
     | {
-     ptyp_desc = Ptyp_arrow (((Labelled _ | Optional _) as lbl), typ1, typ2, _);
+     ptyp_desc = Ptyp_arrow {lbl = (Labelled _ | Optional _) as lbl; arg; ret};
      ptyp_attributes = attrs;
     } ->
       (* Res_core.parse_es6_arrow_type has a workaround that removed an extra arity for the function if the
@@ -39,21 +37,18 @@ let arrow_type ?(max_arity = max_int) ct =
          When this case is encountered we add that missing arity so the arrow is printed properly.
       *)
       let arity =
-        match typ1 with
+        match arg with
         | {ptyp_desc = Ptyp_any; ptyp_attributes = attrs1}
           when has_as_attr attrs1 ->
           arity
         | _ -> arity - 1
       in
-      let arg = (attrs, lbl, typ1) in
-      process attrs_before (arg :: acc) typ2 arity
+      let arg = (attrs, lbl, arg) in
+      process attrs_before (arg :: acc) ret arity
     | typ -> (attrs_before, List.rev acc, typ)
   in
   match ct with
-  | {
-      ptyp_desc = Ptyp_arrow (Nolabel, _typ1, _typ2, _);
-      ptyp_attributes = attrs1;
-    } as typ ->
+  | {ptyp_desc = Ptyp_arrow {lbl = Nolabel}; ptyp_attributes = attrs1} as typ ->
     process attrs1 [] {typ with ptyp_attributes = []} max_arity
   | typ -> process [] [] typ max_arity